--- GettingStarted_2017.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26020.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GettingStarted_2017", "GettingStarted_2017.csproj", "{98D10E89-609D-4E7F-BE7F-87FD125DAC73}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{98D10E89-609D-4E7F-BE7F-87FD125DAC73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{98D10E89-609D-4E7F-BE7F-87FD125DAC73}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{98D10E89-609D-4E7F-BE7F-87FD125DAC73}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{98D10E89-609D-4E7F-BE7F-87FD125DAC73}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {8349719F-68AF-4747-8501-283FB0DBDA83}
-	EndGlobalSection
-EndGlobal
