--- app.config
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<configuration>
-  <runtime>
-    <assemblyBinding xmlns="urn:schemas-microsoft-com:asm.v1">
-      <dependentAssembly>
-        <assemblyIdentity name="Syncfusion.DataSource.WinForms" publicKeyToken="3d67ed1f87d44c89" culture="neutral" />
-        <bindingRedirect oldVersion="0.0.0.0-19.3460.1.1" newVersion="19.3460.1.1" />
-      </dependentAssembly>
-    </assemblyBinding>
-  </runtime>
-</configuration>
