--- packages.config
@@ -1,10 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Syncfusion.Core.WinForms" version="19.3.0.47" targetFramework="net46" />
-  <package id="Syncfusion.DataSource.WinForms" version="19.3.1.2" targetFramework="net46" />
-  <package id="Syncfusion.GridCommon.WinForms" version="19.3.0.47" targetFramework="net46" />
-  <package id="Syncfusion.Licensing" version="19.3.0.47" targetFramework="net46" />
-  <package id="Syncfusion.Office2016Theme.WinForms" version="19.3.0.47" targetFramework="net46" />
-  <package id="Syncfusion.SfListView.WinForms" version="19.3.0.47" targetFramework="net46" />
-  <package id="Syncfusion.Shared.Base" version="19.3.0.47" targetFramework="net46" />
-</packages>
